common resolver utility and shape passed to link target as well
