+++ src/test/java/com/dia/ismdtoolbackend/service/impl/WorkingCopyDeviationServiceImplTest.java
@@ -1,15 +1,11 @@
 package com.dia.ismdtoolbackend.service.impl;
 
 import com.dia.ismdtoolbackend.client.NkdSparqlClient;
-import com.dia.ismdtoolbackend.controller.dto.NkdConceptRefDto;
-import com.dia.ismdtoolbackend.controller.dto.ResolvedConceptDto;
 import com.dia.ismdtoolbackend.entity.ConceptMetadataEntity;
 import com.dia.ismdtoolbackend.enums.SnapshotOrigin;
 import com.dia.ismdtoolbackend.models.OntologyDetailModel.ConceptDetailModel;
 import com.dia.ismdtoolbackend.models.concept.PublishedConceptDeviationModel;
 import com.dia.ismdtoolbackend.models.concept.PublishedConceptDeviationModel.DeviationStatus;
-import com.dia.ismdtoolbackend.models.rpp.RppAgenda;
-import com.dia.ismdtoolbackend.models.rpp.RppIsvs;
 import com.dia.ismdtoolbackend.repository.ConceptMetadataRepository;
 import com.dia.ismdtoolbackend.repository.JenaTDB2Repository;
 import com.dia.ismdtoolbackend.utility.detail.OntologyDetailExtractor;
@@ -23,12 +19,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.junit.jupiter.MockitoSettings;
 import org.mockito.quality.Strictness;
 
-import java.util.Map;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -51,8 +45,7 @@ class WorkingCopyDeviationServiceImplTest {
     @Mock private OntologyDetailExtractor detailExtractor;
     @Mock private ConceptDeviationComparator conceptDeviationComparator;
     @Mock private NkdSparqlClient nkdSparqlClient;
-    @Mock private ReferencedConceptResolutionEngine resolutionEngine;
-    @Mock private com.dia.ismdtoolbackend.service.rpp.RppSnapshotHolder rppSnapshotHolder;
+    @Mock private DeviationResolutionEnricher deviationEnricher;
 
     private WorkingCopyDeviationServiceImpl service;
 
@@ -61,7 +54,7 @@ class WorkingCopyDeviationServiceImplTest {
     private WorkingCopyDeviationServiceImpl newService(WorkingCopyDeviationServiceImpl self) {
         return new WorkingCopyDeviationServiceImpl(conceptMetadataRepository, jenaTDB2Repository,
                 detailExtractor, conceptDeviationComparator, nkdSparqlClient,
-                resolutionEngine, rppSnapshotHolder, self);
+                deviationEnricher, self);
     }
 
     @BeforeEach
@@ -122,61 +115,20 @@ class WorkingCopyDeviationServiceImplTest {
     }
 
     @Test
-    void deviation_enrichesResolvedCharacteristics_bothDiffSides() {
-        String domainLocal = "https://slovník.gov.cz/a/pojem/mistni-domena";
-        String domainPublished = "https://slovník.gov.cz/a/pojem/nkd-domena";
-        String rangeConcept = "https://slovník.gov.cz/a/pojem/cilova-trida";
-        String agendaIri = "https://rpp/agenda/A123";
-        String aisIri = "https://rpp/isvs/S456";
-
+    void deviation_enrichesTheResult() {
+        // Resolution of concept/RPP refs is delegated to DeviationResolutionEnricher (own test); here we
+        // only pin that deviationFor runs it on the comparator's output before returning.
         ConceptDetailModel published = ConceptDetailModel.builder().iri(IRI).build();
         when(nkdSparqlClient.fetchPublishedConcept(IRI)).thenReturn(Optional.of(published));
-
-        // The comparator's output carries the diffs; the service resolves the IRIs on top.
         PublishedConceptDeviationModel deviation = PublishedConceptDeviationModel.builder()
-                .status(DeviationStatus.HAS_DEVIATIONS)
-                .source(NkdConceptRefDto.builder().iri(IRI).build())
-                .domain(diff(domainLocal, domainPublished))
-                .range(diff("xsd:string", rangeConcept))      // one datatype side, one concept side
-                .agenda(diff(agendaIri, agendaIri))
-                .ais(diff(aisIri, null))
-                .build();
+                .status(DeviationStatus.HAS_DEVIATIONS).build();
         when(conceptDeviationComparator.compareConceptDetails(local, published, SnapshotOrigin.WORKING_COPY, IRI))
                 .thenReturn(deviation);
 
-        // Concept resolution covers source + both domains + the concept-typed range value.
-        when(resolutionEngine.resolveAll(anyList())).thenReturn(Map.of(
-                IRI, resolvedConcept(IRI),
-                domainLocal, resolvedConcept(domainLocal),
-                domainPublished, resolvedConcept(domainPublished),
-                rangeConcept, resolvedConcept(rangeConcept)));
-        RppAgenda agenda = new com.dia.ismdtoolbackend.models.rpp.RppAgenda();
-        agenda.setIri(agendaIri);
-        when(rppSnapshotHolder.findAgendaByIri(agendaIri)).thenReturn(Optional.of(agenda));
-        com.dia.ismdtoolbackend.models.rpp.RppIsvs isvs = new RppIsvs();
-        isvs.setIri(aisIri);
-        when(rppSnapshotHolder.findIsvsByIri(aisIri)).thenReturn(Optional.of(isvs));
-
         PublishedConceptDeviationModel result = service.deviationFor(IRI);
 
-        // Concept refs: source + both domain sides + the concept-typed range side, keyed by IRI.
-        assertThat(result.getReferencedConceptsResolved())
-                .containsKeys(IRI, domainLocal, domainPublished, rangeConcept);
-        // The datatype side of range resolves to a DataTypeDto, not a concept.
-        assertThat(result.getRangeResolved()).containsKey("xsd:string");
-        assertThat(result.getReferencedConceptsResolved()).doesNotContainKey("xsd:string");
-        // RPP maps, keyed by IRI, only the sides that were present.
-        assertThat(result.getAgendaResolved()).containsKey(agendaIri);
-        assertThat(result.getAisResolved()).containsKey(aisIri);
-    }
-
-    private static PublishedConceptDeviationModel.PropertyDeviation<String> diff(String local, String published) {
-        return PublishedConceptDeviationModel.PropertyDeviation.<String>builder()
-                .localValue(local).publishedValue(published).isDifferent(true).build();
-    }
-
-    private static ResolvedConceptDto resolvedConcept(String iri) {
-        return ResolvedConceptDto.builder().iri(iri).build();
+        assertThat(result).isSameAs(deviation);
+        verify(deviationEnricher).enrich(deviation);
     }
 
     @Test
